core: Fix clang build

+++ src/core/frontend/applets/controller.cpp
@@ -30,7 +30,7 @@ void DefaultControllerApplet::ReconfigureControllers(std::function<void()> callb
     auto& players = Settings::values.players;
 
     const std::size_t min_supported_players =
-        parameters.enable_single_mode ? 1 : parameters.min_players;
+        parameters.enable_single_mode ? 1 : static_cast<std::size_t>(parameters.min_players);
 
     // Disconnect Handheld first.
     npad.DisconnectNPadAtIndex(8);
